LbsprSimulation class

it generates two dataframes containing information about the estimated vs real catch as well as the estimated unfished vs fished population. It does not apply any depletion threshold for the unfished population.

Value

A list of dataframes with the expected catch and fished/unfished population

Public fields

biology_params

list with the biological parameters of the analysed species

estimates

dataframe with estimates from LBSPR model

catch_data

list of long and wide dataframes

Methods

Public methods

LbsprSimulation$new()

LbsprSimulation$run()

LbsprSimulation$clone()

Method new()

Initialise the class

Usage

LbsprSimulation$new(biology_params, estimates, catch_data)

Arguments

biology_params

list with the biological parameters of the analysed species

estimates

dataframe with estimates from LBSPR model

catch_data

list of long and wide dataframes

Method run()

Generates the model fitting information (relative catch at length and relative fished and unfished population at length)

Usage

LbsprSimulation$run(verbose = TRUE)

Method clone()

The objects of this class are cloneable with this method.

Usage

LbsprSimulation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.