LimeLbms class

it wraps the LIME algorithm providing a common and similar interface to other lbms methods

Value

a list with most LHT-related parameters that LIME needs to run

a list with estimation and simulation results. For the latter, dataframes with the relative expected catch at length, actual catch, relative fished and un-fished populations are returned. While catch is relative to lengths, population details are to ages given that LIME is an age-structured algorithm. All vectors are standarised by dividing them by their total sum.

Super class

lbmstoolbox::Lbms -> LimeLbms

Methods

Public methods

LimeLbms$new()

LimeLbms$prepare_catch_data()

LimeLbms$build_lht_context()

LimeLbms$run()

LimeLbms$clone()

Inherited methods

lbmstoolbox::Lbms$df_to_unname_matrix()

lbmstoolbox::Lbms$transpose()

Method new()

Usage

LimeLbms$new(biological_params, explotation_params, catch_data)

Method prepare_catch_data()

Usage

LimeLbms$prepare_catch_data(data)

Method build_lht_context()

Wrapper over LIME's create_lht_list function that already calculate the starting SL50 AND SL95.

Usage

LimeLbms$build_lht_context(extra_lht_details = NULL)

Method run()

Run LIME algorithm

Usage

LimeLbms$run(lht_context, extra_run_context = NULL, simul = TRUE)

Method clone()

The objects of this class are cloneable with this method.

Usage

LimeLbms$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.